requirements.txt
dj-database-url==0.5.0
Django==3.2
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2020.1
whitenoise==5.2.0
